Rename exonum_service to exonum_interface
