+++ exonum/src/runtime/rust/tests.rs
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 use exonum_crypto::{Hash, PublicKey, PUBLIC_KEY_LENGTH};
-use exonum_derive::exonum_service;
+use exonum_derive::exonum_interface;
 use exonum_merkledb::{BinaryValue, Entry, Fork, Snapshot};
 use exonum_proto::ProtobufConvert;
 use futures::{sync::mpsc, Future};
@@ -222,7 +222,7 @@ struct TxB {
     value: u64,
 }
 
-#[exonum_service(crate = "crate")]
+#[exonum_interface(crate = "crate")]
 trait TestService {
     fn method_a(&self, context: CallContext<'_>, arg: TxA) -> Result<(), ExecutionError>;
     fn method_b(&self, context: CallContext<'_>, arg: TxB) -> Result<(), ExecutionError>;
