+++ services/supervisor/src/transactions.rs
@@ -28,7 +28,7 @@ use super::{
 };
 
 /// Supervisor service transactions.
-#[exonum_service]
+#[exonum_interface]
 pub trait SupervisorInterface {
     /// Requests artifact deploy.
     ///
