+++ test-suite/testkit/src/builder.rs
@@ -80,7 +80,7 @@ use crate::{ApiNotifierChannel, TestKit, TestNetwork};
 /// # Example
 ///
 /// ```
-/// # use exonum_derive::{exonum_service, ServiceFactory};
+/// # use exonum_derive::{exonum_interface, ServiceFactory};
 /// # use exonum_merkledb::Snapshot;
 /// # use exonum_testkit::TestKitBuilder;
 /// # use exonum::{
@@ -104,7 +104,7 @@ use crate::{ApiNotifierChannel, TestKit, TestNetwork};
 /// #     fn state_hash(&self, _: InstanceDescriptor, _: &dyn Snapshot) -> Vec<Hash> { vec![] }
 /// # }
 /// #
-/// # #[exonum_service]
+/// # #[exonum_interface]
 /// # pub trait ExampleInterface {}
 /// #
 /// # impl ExampleInterface for ExampleService {}
