+++ test-suite/testkit/tests/counter/counter.rs
@@ -28,7 +28,9 @@ use exonum::{
         InstanceDescriptor, InstanceId,
     },
 };
-use exonum_derive::{exonum_service, BinaryValue, IntoExecutionError, ObjectHash, ServiceFactory};
+use exonum_derive::{
+    exonum_interface, BinaryValue, IntoExecutionError, ObjectHash, ServiceFactory,
+};
 use exonum_merkledb::{Entry, IndexAccess, ObjectHash, Snapshot};
 use exonum_proto::ProtobufConvert;
 use futures::{Future, IntoFuture};
@@ -103,7 +105,7 @@ pub enum Error {
     AddingZero = 0,
 }
 
-#[exonum_service]
+#[exonum_interface]
 pub trait CounterServiceInterface {
     // This method purposely does not check counter overflow in order to test
     // behavior of panicking transactions.
